refactor + add Makefile
